requirements.txt
pandas==2.2.2
openpyxl==3.1.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
